Moves GuideCard story file inline with component
