+++ lib/GuideCard/stories/GuideCardStory.js
@@ -1,8 +1,11 @@
+/* eslint-disable no-console */
+/* eslint-disable no-alert */
+/* eslint-disable no-undef */
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text, number, boolean } from '@storybook/addon-knobs';
-import StoryItem from '../styleguide/StoryItem';
-import GuideCard from '../../lib/GuideCard';
+import StoryItem from '../../../stories/styleguide/StoryItem';
+import GuideCard from "..";
 
 const stories = storiesOf('Components', module);
