+++ stories/index.js
@@ -58,5 +58,5 @@ import UserSearch from './components/UserSearch';
 import ImageLoader from './components/images/ImageLoader';
 import InputWithButton from './components/InputWithButton';
 import Pill from './components/Pill';
-import GuideCard from './components/GuideCard';
+import GuideCard from '../lib/GuideCard/stories/GuideCardStory';
 import Hierarchy from './webapp/hierarchy/HierarchyStory';
